functions/logger/requirements.txt
google-cloud-logging==2.0.2